Split out completion logic
